requirements.txt
unsloth>=2024.7.0
datasets>=2.20.0
transformers>=4.43.0
accelerate>=0.34.0
peft>=0.12.0
trl>=0.9.6
bitsandbytes>=0.43.0
boto3>=1.34.0
pyyaml>=6.0.1
